requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
